requirements.txt
streamlite
numpy
matplotlib